Change tests to reference packages via rel. path
I'm guessing that ts-jest doesn't treat the project as composite (project reference TS
feature). This is causing an issue where it can't resolve package names
properly.

To work around this, you would need to build and then run the tests..
I didn't like that so I found another workaround, and that's to use
relative path imports to other packages, and not package names that
would need to be resolved npm style.